Fix for "CreateProcess error=206, The filename or extension is too long"
This project did not build for me without this addition, running Android Studio on Windows 8.1. This change adds a plugin to the project's build.gradle script that corrects a build error on Windows arising from long Java classpaths. The error is "CreateProcess error=206, The filename or extension is too long". This error may not occur on all computers running Windows, and there are probably other ways to solve the problem, but this one has the benefits of being a small change and requiring no action on the part of the student of Android programming.